requirements.txt
appdirs==1.4.3
backports.shutil-get-terminal-size==1.0.0
decorator==4.0.11
Django==1.11.1
djangorestframework==3.6.3
enum34==1.1.6
ipython==5.3.0
ipython-genutils==0.2.0
packaging==16.8
pathlib2==2.2.1
pexpect==4.2.1
pickleshare==0.7.4
prompt-toolkit==1.0.14
ptyprocess==0.5.1
Pygments==2.2.0
pyparsing==2.2.0
pytz==2017.2
scandir==1.5
simplegeneric==0.8.1
six==1.10.0
traitlets==4.3.2
wcwidth==0.1.7
httpie==0.9.9
